requirements.txt
asgiref==3.2.10
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
djangorestframework==3.11.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.5
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.2.0
